+++ Sail-a-dex/Plugin.cs
@@ -11,19 +11,26 @@ namespace sailadex
 {
     [BepInPlugin(PLUGIN_GUID, PLUGIN_NAME, PLUGIN_VERSION)]
     [BepInDependency(SMH_GUID, SMH_VERSION)]
+    [BepInDependency(MODSAVEBACKUPS_GUID, MODSAVEBACKUPS_VERSION)]
     [BepInDependency(PASSAGEDUDE_GUID, BepInDependency.DependencyFlags.SoftDependency)]
+    [BepInDependency(RANDOMENCOUNTERS_GUID, BepInDependency.DependencyFlags.SoftDependency)]
     public class Plugin : BaseUnityPlugin
     {
         public const string PLUGIN_GUID = "com.raddude82.sailadex";
         public const string PLUGIN_NAME = "Sail-A-Dex";
-        public const string PLUGIN_VERSION = "1.2.0";
+        public const string PLUGIN_VERSION = "1.3.0";
 
         public const string SMH_GUID = "com.app24.sailwindmoddinghelper";
         public const string SMH_VERSION = "2.0.3";
+        public const string MODSAVEBACKUPS_GUID = "com.raddude82.modsavebackups";
+        public const string MODSAVEBACKUPS_VERSION = "1.0.1";
         public const string PASSAGEDUDE_GUID = "pr0skynesis.passagedude";
+        public const string RANDOMENCOUNTERS_GUID = "com.raddude82.randomencounters";        
 
         internal static Plugin instance;
         internal static ManualLogSource logger;
+        internal static Harmony harmony;
+
         internal static ConfigEntry<bool> fishNamesHidden;
         internal static ConfigEntry<bool> portNamesHidden;
         internal static ConfigEntry<bool> fishCaughtUIEnabled;
@@ -31,14 +38,13 @@ namespace sailadex
         internal static ConfigEntry<bool> statsUIEnabled;
         internal static ConfigEntry<bool> notificationsEnabled;
         internal static ConfigEntry<float> notificationSoundVolume;
-        internal static ConfigEntry<string> updateMilesSailed;
-
-        internal static Harmony harmony;
+        internal static ConfigEntry<string> updateMilesSailed;        
 
         private void Awake()
         {
             instance = this;
             logger = Logger;
+
             fishNamesHidden = Config.Bind("Settings", "Hide Fish Names Before Caught", true, "true = fish names will be hidden before being caught for the first time.");
             portNamesHidden = Config.Bind("Settings", "Hide Port Names Before Visited", false, "true = port names will be hidden before visited for the first time.");
             fishCaughtUIEnabled = Config.Bind("Settings", "Enable Fish Caught UI", true, "true = UI for how many fish you caught will be enabled. Setting to false, continuing a game where previously enabled, and then saving will erase all previous fish caught progress.");
@@ -49,6 +55,7 @@ namespace sailadex
             updateMilesSailed = Config.Bind("Settings", "Miles Sailed Updates", "moored", new ConfigDescription("Miles sailed text will be updated once moored, going to sleeping or moored, or in real time.", new AcceptableValueList<string>("moored", "sleep", "realtime")) );
 
             harmony = Harmony.CreateAndPatchAll(Assembly.GetExecutingAssembly(), PLUGIN_GUID);
+
             foreach (var plugin in Chainloader.PluginInfos)
             {
                 var metadata = plugin.Value.Metadata;
@@ -57,8 +64,15 @@ namespace sailadex
                     logger.LogInfo($"{PASSAGEDUDE_GUID} found");
                     PassageDude.PatchMod();
                 }
-            }
+                if (metadata.GUID.Equals(RANDOMENCOUNTERS_GUID))
+                {
+                    logger.LogInfo($"{RANDOMENCOUNTERS_GUID} found");
+                    RandomEncounters.pluginInstance = plugin.Value.Instance;
+                    RandomEncounters.PatchMod();
+                }
+            }            
         }
+
         private void OnDestroy()
         {
             logger.LogInfo($"Destroying and unpatching {PLUGIN_GUID}");